+++ src/body.rs
@@ -1,8 +1,7 @@
 use std::fmt::{self, Write};
 
-use block::Block;
-use formatter::Formatter;
-
+use crate::block::Block;
+use crate::formatter::Formatter;
 
 #[derive(Debug, Clone)]
 pub enum Body {
@@ -10,12 +9,11 @@ pub enum Body {
     Block(Block),
 }
 
-
 impl Body {
-    pub fn fmt(&self, fmt: &mut Formatter) -> fmt::Result {
-        match *self {
-            Body::String(ref s) => write!(fmt, "{}\n", s),
-            Body::Block(ref b) => b.fmt(fmt),
+    pub fn fmt(&self, fmt: &mut Formatter<'_>) -> fmt::Result {
+        match &self {
+            Body::String(s) => write!(fmt, "{}\n", s),
+            Body::Block(b) => b.fmt(fmt),
         }
     }
 }