+++ src/scope.rs
@@ -1,22 +1,18 @@
-extern crate indexmap;
-
-
 use std::fmt::{self, Write};
 
-use self::indexmap::IndexMap;
+use indexmap::IndexMap;
 
-use docs::Docs;
-use formatter::Formatter;
-use function::Function;
-use import::Import;
-use item::Item;
-use module::Module;
+use crate::docs::Docs;
+use crate::formatter::Formatter;
+use crate::function::Function;
+use crate::import::Import;
+use crate::item::Item;
+use crate::module::Module;
 
-use r#enum::Enum;
-use r#impl::Impl;
-use r#struct::Struct;
-use r#trait::Trait;
-
+use crate::r#enum::Enum;
+use crate::r#impl::Impl;
+use crate::r#struct::Struct;
+use crate::r#trait::Trait;
 
 /// Defines a scope.
 ///
@@ -32,7 +28,6 @@ pub struct Scope {
     /// Contents of the documentation,
     items: Vec<Item>,
 }
-
 
 impl Scope {
     /// Returns a new scope
@@ -239,7 +234,7 @@ impl Scope {
     }
 
     /// Formats the scope using the given formatter.
-    pub fn fmt(&self, fmt: &mut Formatter) -> fmt::Result {
+    pub fn fmt(&self, fmt: &mut Formatter<'_>) -> fmt::Result {
         self.fmt_imports(fmt)?;
 
         if !self.imports.is_empty() {
@@ -267,7 +262,7 @@ impl Scope {
         Ok(())
     }
 
-    fn fmt_imports(&self, fmt: &mut Formatter) -> fmt::Result {
+    fn fmt_imports(&self, fmt: &mut Formatter<'_>) -> fmt::Result {
         // First, collect all visibilities
         let mut visibilities = vec![];