+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# <future version>
+
+### Changed
+- updated the crate to rust 2018 edition
+
 # 0.1.3 (May 9, 2020)
 
 ### Added